+++ test/suite/index.js
@@ -1,42 +1,29 @@
+'use strict';
+
 const path = require('path');
 const Mocha = require('mocha');
-const glob = require('glob');
+const { glob } = require('glob');
 
-function run() {
-	// Create the mocha test
-	const mocha = new Mocha({
-		ui: 'tdd',
-		color: true
-	});
-
+async function run() {
+	const mocha = new Mocha({ ui: 'tdd', color: true });
 	const testsRoot = path.resolve(__dirname, '..');
-
-	return new Promise((c, e) => {
-		glob('**/**.test.js', { cwd: testsRoot }, (err, files) => {
-			if (err) {
-				return e(err);
-			}
-
-			// Add files to the test suite
-			files.forEach(f => mocha.addFile(path.resolve(testsRoot, f)));
-
-			try {
-				// Run the mocha test
-				mocha.run(failures => {
-					if (failures > 0) {
-						e(new Error(`${failures} tests failed.`));
-					} else {
-						c();
-					}
-				});
-			} catch (err) {
-				console.error(err);
-				e(err);
-			}
-		});
+	const files = await glob('suite/**/*.test.js', { cwd: testsRoot });
+	for (const f of files) {
+		mocha.addFile(path.resolve(testsRoot, f));
+	}
+	return new Promise((resolve, reject) => {
+		try {
+			mocha.run((failures) => {
+				if (failures > 0) {
+					reject(new Error(`${failures} tests failed.`));
+				} else {
+					resolve();
+				}
+			});
+		} catch (err) {
+			reject(err);
+		}
 	});
 }
 
-module.exports = {
-	run
-};
+module.exports = { run };